+++ apps/judicial-system/web/pages/krafa/greinargerd/[id].ts
@@ -1,3 +1,3 @@
-import { StepFour } from '@island.is/judicial-system-web/src/routes/Prosecutor/RestrictionCase/StepFour/StepFour'
+import { PoliceReport } from '@island.is/judicial-system-web/src/routes/Prosecutor/RestrictionCase/PoliceReport/PoliceReport'
 
-export default StepFour
+export default PoliceReport
