+++ apps/judicial-system/web/src/types/index.ts
@@ -14,11 +14,11 @@ export enum Sections {
 }
 
 export enum RestrictionCaseProsecutorSubsections {
-  STEP_ONE = 0,
-  STEP_TWO = 1,
-  STEP_THREE = 2,
-  STEP_FOUR = 3,
-  STEP_FIVE = 4,
+  DEFENDANT = 0,
+  HEARING_ARRANGEMENTS = 1,
+  POLICE_DEMANDS = 2,
+  POLICE_REPORT = 3,
+  CASE_FILES = 4,
   PROSECUTOR_OVERVIEW = 5,
 }
